# openapi/growlab.openapi.yaml
openapi: 3.0.3
info:
  title: GrowLab API
  version: 0.1.0
  description: Local GrowLab REST API source of truth.
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /api/zones:
    get:
      tags: [Zones]
      operationId: listZones
      responses:
        "200":
          description: Zones list.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Zone"
    post:
      tags: [Zones]
      operationId: createZone
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ZoneCreateRequest"
      responses:
        "201":
          description: Zone created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Zone"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/zones/{id}:
    get:
      tags: [Zones]
      operationId: getZone
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Zone detail.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Zone"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Zones]
      operationId: updateZone
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ZoneUpdateRequest"
      responses:
        "200":
          description: Zone updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Zone"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Zones]
      operationId: deleteZone
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "204":
          description: Zone deleted.
        "404":
          $ref: "#/components/responses/NotFound"
  /api/zones/{id}/profiles:
    get:
      tags: [Zones]
      operationId: listZoneProfiles
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Target profiles for a zone.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ZoneProfile"
    post:
      tags: [Zones]
      operationId: createZoneProfile
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ZoneProfileCreateRequest"
      responses:
        "201":
          description: Zone profile created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ZoneProfile"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/plants:
    get:
      tags: [Plants]
      operationId: listPlants
      parameters:
        - name: zoneId
          in: query
          required: false
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Plants list.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Plant"
    post:
      tags: [Plants]
      operationId: createPlant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlantCreateRequest"
      responses:
        "201":
          description: Plant created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plant"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/plants/{id}:
    get:
      tags: [Plants]
      operationId: getPlant
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Plant detail.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plant"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Plants]
      operationId: updatePlant
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlantUpdateRequest"
      responses:
        "200":
          description: Plant updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plant"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Plants]
      operationId: deletePlant
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "204":
          description: Plant deleted.
        "404":
          $ref: "#/components/responses/NotFound"
  /api/plants/{id}/timeline:
    get:
      tags: [Plants]
      operationId: getPlantTimeline
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Plant timeline.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlantTimelineItem"
  /api/plants/{id}/events:
    get:
      tags: [Plants]
      operationId: listPlantEvents
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Plant events.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlantEvent"
    post:
      tags: [Plants]
      operationId: createPlantEvent
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlantEventCreateRequest"
      responses:
        "201":
          description: Plant event created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlantEvent"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/plants/{id}/images:
    get:
      tags: [Plants]
      operationId: listPlantImages
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Plant image metadata list.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlantImage"
    post:
      tags: [Plants]
      operationId: uploadPlantImage
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/PlantImageUploadRequest"
      responses:
        "201":
          description: Plant image uploaded.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlantImage"
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /api/plants/{id}/tasks:
    get:
      tags: [Plants]
      operationId: listPlantTasks
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Manual checklist tasks for a plant.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlantTask"
    post:
      tags: [Plants]
      operationId: createPlantTask
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlantTaskCreateRequest"
      responses:
        "201":
          description: Plant task created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlantTask"
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /api/system/events:
    get:
      tags: [System]
      operationId: listSystemEvents
      responses:
        "200":
          description: Global system events.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SystemEvent"
    post:
      tags: [System]
      operationId: createSystemEvent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SystemEventCreateRequest"
      responses:
        "201":
          description: System event created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemEvent"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/system/alerts:
    get:
      tags: [System]
      operationId: listSystemAlerts
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/SystemAlertStatus"
      responses:
        "200":
          description: System alerts.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SystemAlert"
    post:
      tags: [System]
      operationId: createSystemAlert
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SystemAlertCreateRequest"
      responses:
        "201":
          description: System alert created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemAlert"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/system/alerts/{id}/acknowledge:
    post:
      tags: [System]
      operationId: acknowledgeSystemAlert
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SystemAlertTransitionRequest"
      responses:
        "200":
          description: Alert acknowledged.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemAlert"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/system/alerts/{id}/resolve:
    post:
      tags: [System]
      operationId: resolveSystemAlert
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SystemAlertTransitionRequest"
      responses:
        "200":
          description: Alert resolved.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemAlert"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/devices/{id}:
    get:
      tags: [Devices]
      operationId: getDevice
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Device detail.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    Zone:
      type: object
      required:
        [id, growAreaId, name, slug, environmentType, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        growAreaId:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        environmentType:
          type: string
          example: indoor
        metadata:
          $ref: "#/components/schemas/JsonObject"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ZoneCreateRequest:
      type: object
      required: [growAreaId, name, slug]
      properties:
        growAreaId:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        environmentType:
          type: string
          default: indoor
        metadata:
          $ref: "#/components/schemas/JsonObject"
    ZoneUpdateRequest:
      allOf:
        - $ref: "#/components/schemas/ZoneCreateRequest"
...